requirements.txt
streamlit
langchain
python-dotenv
langchain-google-genai
google-generativeai
embedchain